refactor: mover diretórios /errors e /models para /src
