+++ src/middlewares/verify-jwt.ts
@@ -1,7 +1,7 @@
 import { Request, Response, NextFunction } from "express";
 import * as jwt from 'jsonwebtoken';
 import { JWT_SECRET } from "../env/jwt.env";
-import { UnauthorizedError } from "../core/errors/unauthorized.error";
+import { UnauthorizedError } from "../errors/unauthorized.error";
 import * as authService from "../services/auth.service";
 import { UUID } from "node:crypto";
 import { JwtPayload } from "jsonwebtoken";
