+++ src/models/irrigation.model.ts
