+++ src/models/pivot.model.ts
